modified:   App.tsx 	modified:   pages/pages-router.ts 	deleted:    pages/record/record.tsx 	renamed:    router/Screen2.tsx -> pages/user/Screen2.tsx 	renamed:    router/screen1.tsx -> pages/user/screen1.tsx 	modified:   pages/user/user.tsx 	modified:   router/root.tsx
