+++ pages/pages-router.ts
@@ -1,13 +1,16 @@
 import Home from '@/pages/home/home';
 import Map from '@/pages/map/map';
-import Record from '@/pages/record/record';
 import Share from '@/pages/share/share';
 import User from '@/pages/user/user';
+import { Screen1 } from '@/pages/user/screen1';
+import { Screen2 } from '@/pages/user/Screen2';
+
 
 export {
     Home,
     Map as MapContainer,
-    Record,
     Share,
     User,
+    Screen1,
+    Screen2
 }
