--- pages/record/record.tsx
@@ -1,11 +1,0 @@
-import { Text } from "@rneui/themed";
-
-const Record = () => {
-    return (
-        <Text>
-            Record
-        </Text>
-    );
-}
-
-export default Record;
